+++ src/user/rgen.cpp
@@ -6,7 +6,7 @@
 size_t __stdcall rgen(kiv_os::TRegisters &regs) {
 
 	kiv_os::TProcess_Startup_Info* process_info = reinterpret_cast<kiv_os::TProcess_Startup_Info *>(regs.rdx.r);
-	int pid = regs.rcx.r;
+	int pid = (int)regs.rcx.r;
 	char *arg = process_info->arg;
 
 	kiv_os::THandle std_in = process_info->std_in;
